KLEIP-3
This commit adds the KLEIP-3 proposal which aims to introduce a temporary block on GP distribution decisions among battle participants. The block will last for 3 months, during which GP can still be used for staking or exchanges that benefit the community. This measure ensures decisions are made thoughtfully and with communal agreement, enhancing overall token value.